refactor(transformService): move request transforms into their respective files
